+++ src/odr/document_element.cpp
@@ -1,6 +1,5 @@
 #include <odr/document_element.hpp>
 
-#include <odr/document_path.hpp>
 #include <odr/exceptions.hpp>
 #include <odr/file.hpp>
 #include <odr/style.hpp>
@@ -141,18 +140,6 @@ bool Element::is_self_locatable() const {
 
 bool Element::is_editable() const {
   return exists_() ? m_adapter->element_is_editable(m_identifier) : false;
-}
-
-DocumentPath Element::document_path() const {
-  return exists_() ? m_adapter->element_document_path(m_identifier)
-                   : DocumentPath();
-}
-
-Element Element::navigate_path(const DocumentPath &path) const {
-  return exists_()
-             ? Element(m_adapter,
-                       m_adapter->element_navigate_path(m_identifier, path))
-             : Element();
 }
 
 TextRoot Element::as_text_root() const {